f3_c360_notify_results/requirements.txt
aws-psycopg2==1.3.8
boto3==1.38.18
botocore==1.38.18
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.12.0
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
Unidecode==1.4.0
valkey-glide==1.3.5